+++ cimpy/__init__.py
@@ -1,6 +1,6 @@
 from cimpy.cimexport import cim_export
 from cimpy.cimimport import cim_import
-import cimpy.utils
+import cimpy.cimedit
 from cimpy.cimexamples import import_example
 from cimpy.cimexamples import export_example
 from cimpy.cimexamples import addExternalNetworkInjection_example
